面向API使用独立的json序列化和反序列化 面向r2dbc定义独立的转换器

+++ src/main/java/org/huhu/test/platform/converter/TestPlatformRoleLevelToByteConverter.java
@@ -14,7 +14,7 @@ public class TestPlatformRoleLevelToByteConverter implements Converter<TestPlatf
 
     @Override
     public Byte convert(TestPlatformRoleLevel source) {
-        return source.getLevel();
+        return (byte) source.getLevel();
     }
 
 }
